Split deposit.tsx into 2 components
